LayoutTests/fast/images/img-dimensions-styled-expected.txt

Should used styled image dimensions if available.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS image.width is 200
PASS image.height is 200
PASS rect.width is 200
PASS rect.height is 200
PASS successfullyParsed is true

TEST COMPLETE
